requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#


adal==1.2.7
applicationinsights==0.11.10
argcomplete==1.12.3
azure-cli-core==2.25.0
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.16.0
azure-identity==1.6.0
azure-mgmt-consumption==8.0.0
azure-mgmt-containerregistry==8.0.0
azure-mgmt-core==1.2.2
azure-mgmt-dns==8.0.0
azure-mgmt-loganalytics==10.0.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-resource==18.0.0
azure-mgmt-sql==3.0.0
azure-mgmt-subscription==1.0.0
azure-nspkg==3.0.2
azure.loganalytics==0.1.0
azure.mgmt.compute==21.0.0
azure.mgmt.monitor==2.0.0
azure.mgmt.network==19.0.0
azure.mgmt.recoveryservices==1.0.0
azure.mgmt.recoveryservicesbackup==1.0.0
azure.mgmt.storage==18.0.0
bcrypt==3.2.0; python_version >= '3.6'
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.1
colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.3.2
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
cssutils==2.3.0; python_version >= '3.6'
humanfriendly==9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
isodate==0.6.0
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
knack==0.8.2
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown==3.3.4
msal-extensions==0.3.0
msal==1.12.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.1; python_version >= '3.6'
paramiko==2.7.2
pkginfo==1.7.0
portalocker==1.7.1; platform_system != 'Windows'
premailer==3.9.0
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.9.0; python_version >= '3.5'
pyjwt[crypto]==1.7.1; python_version >= '3.6'
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyopenssl==20.0.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests-oauthlib==1.3.0
requests==2.25.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabulate==0.8.9
urllib3[secure]==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
